requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.1.3
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
